+++ frontend/src/app/signup-success/page.tsx
@@ -2,7 +2,7 @@
 
 import { useState, useEffect } from 'react'
 import { useRouter, useSearchParams } from 'next/navigation'
-import { useAuth } from '@/contexts/AuthContext'
+import { useAuthContext as useAuth } from '@/contexts/auth'
 
 // Dynamic rendering を強制する
 export const dynamic = 'force-dynamic'
