+++ frontend/src/components/auth/EmailVerificationBanner.tsx
@@ -2,7 +2,7 @@
 
 import { useState } from 'react'
 import { useRouter } from 'next/navigation'
-import { useAuth } from '@/contexts/AuthContext'
+import { useAuthContext as useAuth } from '@/contexts/auth'
 
 interface EmailVerificationBannerProps {
   email: string
