+++ frontend/src/components/growth-records/GrowthRecordList.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { useState, useEffect, useCallback, useRef } from 'react'
-import { useAuth } from '@/contexts/AuthContext'
+import { useAuthContext as useAuth } from '@/contexts/auth'
 import { useApi } from '@/hooks/useApi'
 import GrowthRecordCard from './GrowthRecordCard'
 import CreateGrowthRecordModal from './CreateGrowthRecordModal'
