+++ frontend/src/components/timeline/TimelinePost.tsx
@@ -1,4 +1,4 @@
-import { useAuth } from '@/contexts/AuthContext'
+import { useAuthContext as useAuth } from '@/contexts/auth'
 import { useImageModal } from '@/contexts/ImageModalContext'
 import { useState } from 'react'
 import Link from 'next/link'
