+++ frontend/src/components/ui/RedirectMessageHandler.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { useEffect } from 'react'
-import { useAuth } from '@/contexts/AuthContext'
+import { useAuthContext as useAuth } from '@/contexts/auth'
 import { useFlash } from '@/contexts/FlashContext'
 
 /**
